+++ kernel/kernel/string.h
@@ -9,6 +9,6 @@ char *strcpy(char *dest, const char *src);
 char *strcat(char *dest, const char *src);
 void memset(void *dest, u8 val, u32 len);
 void memcpy(void *dest, const void *src, u32 len);
-void itoa(char *buf, int base, int d);
+void itoa(char *buf, u64 base, u64 d);
 
 #endif
